rewrote how context switching is done within the scanner
this allows for new functionality: re-use a previously defined context


git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk@3963 7640d186-872a-4c71-b7b2-a5696852cc35

+++ testcode/bftestview2/bftextview2_patcompile.h
@@ -4,10 +4,10 @@
 
 #include "bftextview2.h"
 
-guint16 new_context(Tscantable *st, gchar *symbols, GtkTextTag *contexttag, gboolean autocomplete_case_insens);
+gint16 new_context(Tscantable *st, gchar *symbols, GtkTextTag *contexttag, gboolean autocomplete_case_insens);
 void match_set_nextcontext(Tscantable *st, guint16 matchnum, guint16 nextcontext);
 void match_set_autocomplete(Tscantable *st,gchar *pattern,guint16 context,gchar *autocomplete_append,gchar *reference);
-guint16 add_keyword_to_scanning_table(Tscantable *st, gchar *keyword, gboolean is_regex,gboolean case_insens, GtkTextTag *selftag, guint context, guint nextcontext
+guint16 add_keyword_to_scanning_table(Tscantable *st, gchar *keyword, gboolean is_regex,gboolean case_insens, GtkTextTag *selftag, gint16 context, gint16 nextcontext
 				, gboolean starts_block, gboolean ends_block, guint blockstartpattern
 				, GtkTextTag *blocktag,gboolean add_to_ac, gchar *append_to_ac, gchar *reference);
 void print_DFA(Tscantable *st, char start, char end);
